requirements.txt
pandas==1.5.0
torch==2.3.1
pillow==9.2.0
wildlife_tools==0.0.9